test: 💍 move sample JSON documents to /src/__tests__ folder
