+++ src/json-crdt/codec/compact-binary/__tests__/automated-server.spec.ts
@@ -2,7 +2,7 @@ import {LogicalVectorClock} from '../../../../json-crdt-patch/clock';
 import {Model} from '../../../model';
 import {ServerEncoder} from '../ServerEncoder';
 import {ServerDecoder} from '../ServerDecoder';
-import {documents} from '../../../../util/__tests__/json-documents';
+import {documents} from '../../../../__tests__/json-documents';
 
 for (const {name, json} of documents) {
   test(name, () => {
