+++ src/json-crdt/model/__tests__/Model.automated.set-root.spec.ts
@@ -1,5 +1,5 @@
 import {Model} from '../../model';
-import {documents} from '../../../util/__tests__/json-documents';
+import {documents} from '../../../__tests__/json-documents';
 
 for (const {name, json, only} of documents) {
   (only ? test.only : test)(name, () => {
